Refactor: define primitives separately
This makes the code a bit easier for me to follow, because things have
names and are not as deeply nested in where blocks.
